+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli.Evaluations/Models/ScenarioData.cs
@@ -1,11 +1,6 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Microsoft.Extensions.AI;
 
-namespace Azure.Sdk.Tools.Cli.Evaluations
+namespace Azure.Sdk.Tools.Cli.Evaluations.Models
 {
     public class ScenarioData
     {
